\name{MRG32k5a}
\alias{MRG32k5a}
\title{
  Test function for the package rngOpenMP in sequential mode.
}
\description{
  R level implementation of the generator MRG32k5a using GMP library
  intended for a verification of the C-level functions from the
  package rngOpenMP.
}
\usage{
  MRG32k5a(state, n)
}
\arguments{
  \item{state}{ Numeric vector of length 10 representing the internal
        state of MRG32k5a generator.}
  \item{n}{ Number of random numbers to be generated.}
}
\details{
  The function generates \code{n} random numbers from MRG32k5a generator
  initialized by \code{state}. May be used for comparison with the random
  numbers generated by the C-level functions in rngOpenMP. See the test
  script \code{"rngVerify/tests/testGMP.R"} for a possible use.
}
\value{
  Numeric vector of length \code{n} representing the output of MRG32k5a generator.
}
\author{
  Petr Savicky.
}
\examples{
  library(gmp)
  MRG32k5a(rep(12345, times=10), 5)
  # [1] 0.2581892 0.6479070 0.9535788 0.4342718 0.2002761
}

